Check in the right bitfields file.
The nrf52_bitfields.h file from the SDK was the wrong file to use.
It only contains peripheral register definitions for those devices
common to ann NRF52 devices. The nrf52840_bitfields.h file includes all
the new things in the 52840 chip, like the QSPI controller, which is
what we want.