Add preliminary SPI screen/SD card support
Unfortunately, the SPIM3 bus on the preview nRF52840 board doesn't work, so
I can't test the correct configuration, but for now I have the screen and SD
card running on bus SPIM0. Speed is still restricted to 8MHz.

For now the SPI bus and chip select pins are connected to pins 1 through 8
on GPIO port P1. This may not be the final configuration though. Some
consultation with John will be required. But for now everything works
as expected. (The touch controller is also hooked up for the time being.)