Fix it so that the badge_832 BSP compiles again
The NRF52832 doesn't have the GPIO pins where the screen and SD
card now reside, but we still want to be able to build and run
ChibiOS on the older boards for compatibility testing.

While I'm at it, also check in the updated documentation for
FatFs 0.13b, which I somehow forgot to do previously.