Make a few more cleanups to the QSPI driver
Always use the shortcut command registers for erase.

Also, once we do a read access via XIP, we need to do an access via one
of the shortcut registers again before we can issue custom commands again.
The m25q driver tries to do a dummy read to exit out of XIP mode. To make
this work, qspi_lld_receive() has been modified to look for a read with
a command opcode of 0 and will issue a read via shortcut registers. This
unjams the interface again.