Add support for having a filesystem on the QSPI flash
Since we have additional storage on QSPI, we may as well do something
useful with it. The FatFs support has been extended to have a second volume,
and a diskio driver has been added on top of the ChibiOS flash support.
Since the QSPI is memory mapped, reads should be fairly quick. (In
retrospect, it may make more sense to use DMA for the reads, but further
experimentation will be needed to tell.)

Also, the qspi_lld_stop() routine has been filled in a little more.