Fix a few things with the QSPI FatFs support
One of the #ifdefs in diskio.c was wrong, which broke the QSPI
disk_initialize() method.

In order to use the second driver easily, we need to enable support
for f_chdrive() in FatFs.

Also, the 52832 doesn't have QSPI, so make sure the support for it is
disabled in the badge_832 Makefile so we don't get a linker error.