Do some more QSPI refinements
In the qspi_lld_receive() function, we need to special-case the
FAST_READ command too, in order for the flashRead() API to work
correctly.

Now that this works, switch to using flashRead() in qspi_disk_read() too.
This means FatFs accesses will use DMA instead of a memcpy(). The result
is the same, but using DMA in theory requires fewer CPU cycles.

I think the fastest mechanism overall may be to directly call qspiReceive()
and qspiSend() from drv_qspi_lld.c, but it's hard to benchmark until I get
a production board.