Add some additional synchronization
Add a dsb (data sync barrier) instruction at the end of all the asynchronous
DMA starting functions so that all of the register updates are written out.
This is required with certain C compiler optimization levels to make sure
transfers start before we start looking at the state registers for the
SPI controller.