Switch to using the high speed SPI bus
We now use SPIM3 instead of SPIM0. The bus clock is set for 32MHz,
the fastest it can go. Sadly, we appear to be limited by the throughput
of the SD card.

Note that the pins for the SPI bus have been changed. While the
NRF52840 does have an additional 16-pin port of GPIO pins, you
can only use them for low speed I/O, according to the manual.
Using SPI0 at 8MHz works ok, but using SPI3 at 32MHz does not.