Add code to switch flash chip to SPI mode.
The Macronix serial flash parts default to operating in single-I/O mode.
That means they only use one serial data line. We need to switch the
chip to quad-I/O mode for QSPI operation. This is done by setting a
bit in the status register. This is documented in a Macronix application
note, though not in the chip reference manual (at least, I couldn't find
it). The Nordic QSPI example has code in it for this, but it doesn't
go out of its way to explain why it's there.

This is a non-volatile bit, so we only need to set it once. The Nordic
boards come with the Macronix chip still at factory default settings
so we need to do it the first time we boot up on one of these boards.
This will likely not be needed for production boards which will probably
have a Micron QSPI part.