Update EXT module to support additional GPIO pins
The NRF52840 has a second bank of 16 GPIO pins. The GPIOTE module
supports the second bank by adding an extra bit to the pad selection
field. Bits 6-12 define pin number to use, and bit 13 defines which
port to use (0 or 1). In order to allow the user to specify that a
pin should be on port 1, we need to adjust the EXT_MODE_GPIO_MASK
definition to allow one extra MSB.