Move touch chipsel and screen chipsel pins
The touch controller SPI chip select and screen SPI chip select pins have
been moved to P0.3 and P0.4. This leaves all of port 1 empty (for now).

The reason for this is that it makes it easier to position most of the
prototype screen module pins together on the same header on the
NRF52840 DK development board. The exception is the SD card chip
select, which is on a header by itself at P0.13.

Eventually we may need some of the P1 pins for the I2S controller.
Hopefully they will work well enough for this: the manual says they
should be used for low-speed I/O only, but the MCLK can be as fast
as 16MHz in some cases.

Any buttons can use the port 1 pins.