Add initial I2S audio support
A driver for the Nordic I2S audio controller has been added. This is
intended to work wih the Cirrus Logic CS4344 I2S codec. Currently the
audio sample rate is 15625Hz and audio is monaural. This can be changed
to stereo fairly easily, but it increases the file size.

A script to encode audio using SOX has also been added.

The API is similar to what was in the DC25 badge.

There will likely be some streamlining later.