Fix some bugs related to advertising and chat
When advertising, we must allocate a static buffer to hold our
advertising data instead of allocating it from the heap and then
freeing once advertising is started. The SoftDevice does not copy
the data to a private buffer: instead it expects the advertisement
block to remain unchanged. With the previous code, pointer to the
advertisement block ceased to be valid once we called chHeapFree(),
and the buffer could be re-used for other things, which would
suddenly cause us to send invalid advertising data at seemingly
random times.

Also, we need to implement a queue to handle radio events. Having just
a single bufferf causes us to lose too many advertisement events.

The chat app logic also needs to be updated so that it does not try
to create a connection from both sides at once. If we're in the "central"
role, then we can respond to a GAP connect event by creating an L2CAP
tunnel. We don't also have to do it from the "peripheral" side too.
Having both sides try to do it sometimes results in a connection
failing.