Move the SD card and touch controller to a separate SPI bus
The SD card and touch controller are now on the SPI0 bus, leaving the
display on the fast SPI3 bus. Three more GPIO pins have been allocated
for this.

Also, one of the I2C controllers is now enabled. Two pins have also been
allocated for this. I haven't had a chance to test this yet, but ChibiOS
already has a driver for this peripheral and I have no reason to think it
won't work. (I'll need to drag out the scope to be sure. For now it
doesn't hurt anything to enable this.)