Fix a memory leak in the radio event queue handling
When a app exits, we have to free all the memory allocated for any
radio events left in the queue, not just reset the producer/consumer
indexes. Otherwise we may gradually leak memory. A new function
to flush out the queue has been added which is now called at thread
creation and termination.