Fix a priority order violation assert flagged by ChibiOS
When debug checks and asserts are turned on in chconf.h, ChibiOS will
flag a "priority order violation" error when the system is running. This
seems to be due to the SysTick interrupt priority. Previously it was set
to 1. Changing it to 5 (a lower priority) seems to correct the error.
This is only a problem when the SoftDevice is enabled, but we need it
enabled for the radio to work.

All IRQ priorities are now configurable via the mcuconf.h file. Right
now they're all set to 5.

This took me far too long to track down. :C