Move memory barriers to graphics driver
I don't think the SPI driver really needs to have the memory barrier
instructions in it. I think the write_index() and write_data() functions
are the real problem children. Moving the __DSB() instructions there seems
to make it more reliable. While working on the video player the screen
would fail to initialize sometimes, and it looks like this may have had
to do with the compiler re-arranging the code a little. I *think* it's
better now.