Make video player fully functional
The video player can now play full audio/video files. The audio sample rate
is also now 15625Hz in stereo. Video plays at 13 frames/second.

Getting video and audio to sync was a little tricky. Unfortunately the
Nordic I2S controller only provides only a very limited set of clock
frequencies, which limits what frame rates we can select. John wanted
16KHz. The closest we can get without going too crazy is 15625Hz.
The problem is there's no way to get a whole number of samples per
scanline like we could with the DC25 badge.

The only way to deal with this is to match the video and audio sample
rate as best as we can, and then drop a sample once in a while when
nobody's looking. Using stereo, a sample rate of 15625Hz results in
31250 samples per second. Using a video frame rate of 13 frames/second,
with 120 lines per second, yields 1560 lines/second. That also yields
20.032 samples per scanline. If we round down to 20 samples, that
gives us 31200 samples/second. This is within 50 samples of the desired
rate. To make sure the video and audio stay in sync, we drop every
625th sample. This loses exactly 50 samples.

A new sndskip program has been added to post process raw audio files to
do the sample dropping. Ideally this logic should be merged into the
videomerge program, but math is hard.