Update the video encoding script.
So, the gimmick of occasionally dropping a sample works, but it has a
problem: it creates an occasional click/pop in the audio stream.

I'm not sexy enough to figure out how to work around this (I assume it
would take some kind of smoothing algorithm), so I'm going to take the
easy way out and change the video encoder script so that it simply
re-encodes the audio for a rate of 15600Hz using the sox utility. It's
way smarter about audio than I am. The rate is close enough that the
listener can't really tell the difference and sycnhronization is
right where we want it.