Do some major cleanups of the I2S module
The i2sSamplesSend()/i2sSamplesNext() API has been replaced by just
i2sSamplesPlay()/i2sSamplesStop(). You can just keep resubmitting
samples with i2sSamplesPlay() until you're done, then use i2sSamplesStop()
to turn off the channel. (This is necessary to gate off the SDI clock.)

Also, the i2sSamplesWait() API now no longer needs to rely on a busy
loop. Contrary to what I expected, the I2S controller doesn't interrupt
when the current batch of samples has finished being transfered to the
codec. It interrupt before that, so I had to resort to using a loop
to test the EVENTS_TXPTRUPD register. The problem with this is that
doing it without a sleep hogs the CPU, and doing it with a sleep can
result in sleeping too long unless you jigger the system clock frequency
(which is what I was doing before).

To get around this, I decided to make use of the Nordic PPI module
to connect the EVENTS_TXPTRUPD event to a software interrupt channel.
Now I can generate an interrupt just when the register changes state,
which means I can put the waiting thread to sleep and it will wake
up exactly when the audio samples finish playing without any loops
of mucking about with timers.

Also, I modified the encode_video.sh script a little bit to try and
avoid what I think is some distortion in the audio due to overloading
when we use the "contrast 80" option to sox.