Do a round of UI/sound cleanups
There was a bug in the I2S audio player. When queueing up samples, we
should play however many samples we read from the file, not I2S_SAMPLES.
This didn't appear to be a problem until I tried playing the click
sound effect, which is really short.

The joypad support has been reworked. We can't just call the orchard
event handler routines from the joypad thread: that will cause the
joypad event thread to race with the orchard app thread. In some
cases this race condition can be fatal. Instead we gather the keypress
info and signal an event using the dispatcher.

Playing of the click sound effect has been re-enabled. The click sound
has to be re-encoded using the updated conversion script.

The launcher app needs some rework. It seems pretty brittle when it
comes to undrawing/redrawing the buttons. I think I need to take it
apart and put it back together again.

Also, the credits app from the DC25 badge has been added. It needs to
have the Super Mario tune re-encoded to work. It works much smoother
on the newer hardware.