Fix possible race with gtimer thread
I noticed that sometimes the icons on the launcher would appear to have
some corruption after pressing the up/down arrow buttons on the screen
(via touch).

The problem is that touch events are polled and processed by the gtimer
thread. Button up/down events can cause button appearance to change a
little, i.e. buttons are highlighed when pressed and returned to normal
when released. This means the gtimer thread will do some screen draws
when button up/down events occur.

This also causes minor changes to the display state variables, notably
the clip regions.

Because of reasons, the launcher app doesn't allow the gwin code to draw
the button icons: instead we draw them manually with putImageFile(), which
uses gdisp routines directly.

The gdisp routines also depend a little on the display state.

In order for our private screen drawing routines and the gwin/gtimer
drawing routines to play nice together, they need to be properly
interlocked using the gwin mutex. Otherwise, the gtimer thread and
app thread will both be trying to access the display state and possibly
step on each other.

In this case, the putImageFile() routine was being interrupted by the
higher-priority gtimer thread and display state was being changed
behind its back, causing some icon pixels to be display in the wrong
locations.

To correct this, we use the _gwinDrawStart()/_gwinDrawExit() routines,
which internally acquire/release the gwin mutex, when using putImageFile()
to draw the icon images. This synchronizes us properly with the gtimer
thread and avoids the corruption.

Note that the ui-keyboard module had a similar issue -- I managed to
avoid it before just by fudging the clip region state, but really we
should be taking/giving the mutex there too.