Upgrade S140 SoftDevice to version 6.1.1
Something seems to have changed with the bootstrap code in this version
of the SoftDevice. It gets stuck in a loop during boot, always branching
back to the entry point at 0xA81, unless I patch a couple of locations
at address 0xFF8 and 0xFFC to 0xFFFFFFFF instead of 0. (Strictly speaking,
it seems only necessary to patch offset 0xFF8.) I'm not sure what this is
all about, but aside from that, the SoftDevice seems to work.