Display reset causes at boot
The nRF52 has a reset reason register we can use to determine why
the CPU was reset (soft reset, reset pin assert, watchdog, etc...). We
may as well display this info.

Also, when the SoftDevice encounters a fault, it calls the fault handler,
which currently does nothing. Now we use it to stash the SoftDevice
fault info and display it on boot as well. The SoftDevice will trigger
a fault if it doesn't detect timer or radio events often enough. This
can happen if you pause the system with the debugger. When you resume,
the SoftDevice will detect the problem and reset the system. Now we
can tell when that happens. (Unfortunately I don't know how to decode
the event ID and info codes.)