Add the touch tone dialer app
This is a port of the app from the DC25 badge, with fixups for using the
nRF52 IS2 controller instead of a DAC.

Note: the dialer app turns off the radio and then turns it back on again
on exit. When the SoftDevice is enabled, there seems to be a fluctuation
in the I2S clock(s) which causes a "wobble" in the tones. This doesn't
seem to affect them being decoded, but it sounds weird, so we work around
it for now.