Add an initial port of the music player app
The bar graph is not yet enabled. I need to figure out how to implement
that given that we now have stereo sample data in our sound files.

Also, fixed a bug in the scroll and music apps: the geventDetachSource()
function clears the mouse event data, so we have to evaluate whether
to exit with 0 or -1 status before calling it, not after. This bug would
prevent the exit click sound from playing in some cases.