Add an sd card build directory
We're likely not going to keep the same icons for launcher screen layout,
but for now I'd like to have a basic build infrastructure for the SD card
assets. This is basically a copy of what we had for the DC25 badge, except
I added rules to build the videos on the fly. Only the original video
files should be checked into the repo (ditto for images and sound effects).

The Makefile will likely need some trimming. For example we probably
won't need the image flipping rule anymore.

When checking in asset files, stick to using an 8.3 filename convention
since the SD card uses a FAT filesystem.