Move the stubs to assembly code
Newer versions of GCC emit warnings due to us aliasing various different
functions to _port_irq_epilogue() (it knows the types don't match). Trying
to prevent this in C is messy, but we can just as easily move the stubs
to the entry.S file and achieve the same effect.