Avoid byte swapping pixels in driver
Fix another thing that's been bugging me for a while: I don't want to
have to byte-swap pixels in the ILI9341 display driver. I added this
before because the nRF52840 SPI controller doesn't have a way to perform
internal byte-swapping. With the KW01 chip in the DC25 badge, you could
specify that data was to be handled 16 bits at a time, and it would send
the pixels out onto the SPI bus in the right order, but tge nRF52840
controller doesn't do that, so it sends them out a byte at a time.
Unfortunately this causes the pixel data written to the screen to be
byte-swapped, so the colors come out wrong. The __builtin_bswap16() I
put into the driver was a workaround for this.

But that means needing to swap data for all graphics writes (except in
the video player), which I want to avoid. Instead the byte swapping is
now done in the uGFX color generation macros themselves. Most of the
time this should result in the colors being byte-swapped at compile
time instead of runtime, so the performance overhead should be negligible.

This also means that image files on the SD card can be written directly
in the "native" pixel format, and drawing them no longer requires any
byte swapping. To achieve this, the convert_to_rgb.sh script now uses
rgb565be as the image conversion type instead of just rgb565. The
encode_video.sh script was already doing this. Now everything is consistent.