Fix the high frequency clock initialization
It turns out the ChibiOS hal_lld.c code was setting the chip to use the
internal RC oscillator for the 64MHz CPU clock. This is why audio from
the I2S chip had a warble it whenever the radio was enabled: the extra
CPU load from the radio was causing oscillator drift. With the external
crystal and the HXCO enabled, it's perfectly steady.

The dialer app no longer disables the radio when its running now.