Add support for powering up/down audio amp
This doesn't exist on the DK board, but the LM4880 audio amp on the
real board can be put into low power mode via GPIO pin P0.02 (high
to power off, low to power on). This disconnects the bias supply to
the amplifier transistors. Consumption should be 7uA when we're not
playing sound. The I2S module now toggles this line when playing a
tune, and an API is now available for other consumers of the I2S
controller (video player, music player, DTMF dialer), to manipulate
it as well.