Avoid initializing SPI bus 4 on broken preview chips
The rev AAAA nRF52840 chips don't have a working high speed SPI bus
controller, and we'll get stuck trying to boot on one of those when
we try to enable the display. I have one preview DK board that I use
for radio debugging and I'd like to avoid this, so if we detect
the preview silicon, we skip the graphics setup.