Make some updates to the BLE GAP code
We now support advertising our name via scan responses. Scan responses
can be sent by devices which identify as "scannable." After a peer
receives an advertising broadcast for a scannable device, it can send
a scan request to obtain an additional 31 bytes of identifying payload.
This is handy because squeezing a nice name into 31 bytes along with
UUID and other info can be tricky.

It turns out that in order to reliably receive scan responses, we have
to request active scanning, and we also have to be sure to call
sd_ble_gap_scan_start() each time we receive a GAP advertisement
report. This was not made quite clear by the API documentation, but
that's what the Nordic SDK code does. Once we get an advertisement,
the SoftDevice code will send out a scan request packet, and it seems
you must do a sd_ble_gap_scan_start() again quickly in order to receive
the scan response.

This allows us to have longer device names. Instead of DC27 IDES,
we now use DC27 Ides Of DEF CON. This may be changed to the user's
name once we have a configuration app.

The chat app has also been updated. It will now display either a MAC
address or a name, if the peer device has one. Eventually this will need
to be restricted to only display other badges by filtering on our vendor ID.