Fix an issue with the DTMF dialer app
When playing tones, we need to temporarily boost the priority of the
DTMF dialer thread. The BLE SoftDevice event handler thread has the
highest priority, and if it interrupts the tone dialer, it creates
small gaps when playing some tones which sound like radio interference
but really aren't. (Well, not the same kind of radio interference we
had witht he previous board anyway.)

The dialer app does have code to boost priority, but it's assuming
the old badge priority scheme. The SoftDevice thread runs at HIGHPRIO,
which is the highest available. To address this, we knock the SD thread
priority down a few notches to allow the dialer app room to outmuscle
it. This does not seem to adversely affect handling radio events (and
you can't really do anything with the radio while the dialer app is
running anyway).