Add some enhancements to the video player module.
Like the DC25 badge, add a function to allow playing un-expanded video
clips at arbitrary places on screen. (We never used this feature last
time, but maybe this year will be different.)

Also, we can suspend the current thread to wait for the display SPI bus
to become ready instead of using a while() loop.