Switch to using the newlib malloc(), free() and printf() routines.
Previously we were relying on the chprintf(), chHeapAlloc() and chHeapFree()
functions in ChibiOS. As of now, we can use the slightly more standard
APIs provided by newlib. We need to disable the ChibiOS heap allocation
routines because otherwise we'll have dueling malloc()s (both newlib and
ChibiOS will be fighting over the same heap storage).

There should be little or no difference in behavior, except for the
following things:

- use the <stdio.h> and <stdlib.h> headers instead of chprintf.h
- use printf(), malloc() and free() instead of chprintf(), chHeapAlloc()
  and chHeapFree()
- You no longer need to use /r/n when doing a printf() (though if you
  still do, it's harmless)

A new "mem" shell command is included since the built-in one no longer
works.