Fix possible crash with some versions of GCC
The Nordic SDK defines NRF_SOC_EVTS as an enum. This describes the events
that can be returned when making a SOC system call to the SoftDevice,
particularly when performing a flash erase or write. Here's something
you need to always remember: while most of the time people assume that
enums will be represented by an int, the truth is that the C specification
says that they're implementation-defined. The compiler could choose to
represent them internally in any form that it wants.

John builds the badge code with this version of the GNU ARM Embedded
toolchain:

GCC 6.3.1 20170215 (release) [ARM/embedded-6-branch revision 245512]

When you build the code with this compiler, you get this (dumped from
GDB using x/xb):

0x2000535c <peer_pk+59>:        0x00
0x2000535d <peer_pk+60>:        0x00
0x2000535e <peer_pk+61>:        0x00
0x2000535f <peer_pk+62>:        0x00
0x20005360 <peer_pk+63>:        0x00
0x20005361 <soc_evt>:   0x00
0x20005362:     0x00
0x20005363:     0x00
0x20005364 <GDISP>:     0x98
0x20005365 <GDISP+1>:   0x80
0x20005366 <GDISP+2>:   0x00
0x20005367 <GDISP+3>:   0x20

Note that soc_evt here is represented by a single byte. There are two
pad bytes after it, followed by the GDISP pointer. The problem here is
that we obtain the event from the SoftDevice like this:

                        r2 = sd_evt_get ((uint32_t *)&soc_evt);

This is because sd_evt_get() expects a uint32_t * pointer. But that
requires 4 bytes. What ends up happening is that when we store the
event in soc_evt, we overwrite the last byte of the GDISP pointer,
so 0x20008098 becomes 0x20008000. After that, apps that use the uGFX
library will crash.

To correct this, soc_evt and flash_evt are now explicitly defined as
uint32_t variables. Technically you shouldn't compare these to enum
values, but that is less morally objectionable than crashing the system.