Add a submarine ping sound effect
Orchard UI events (screen touch, button push) now play a submarine sonar
ping effect. It remains to be seen if we stick with this or not.

For other effects the click effect is retained. For things like the
chat keypad, the sonar ping is too long and too annoying to hear on
every keypush.