Continue work on GATTC client
I'm still learning how all this works. Probing remote devices for their
services and UUIDs is... messy. But I think I'm starting to get it.
It's not done yet, but I want to stop here and see if John has anything
to merge before I go further.

I made one modification to the userconfig code: getConfig() will return
a pointer to the flash userconfig region if it's used before the userconf
module has been initialized. This is needed to be able to read the unlocks
field of the userconfig from flash during the BLE startup.