Implement GATTC read/write APIs
These APIs allow us to read/write characteristics values on a peer badge.
The handle values are the values stored in pw_handle, ul_handle and so
on, which are generated when the GATT server module is initialized. All
badges will implement the same services, so the handle values will all
be the same. This saves us from having to discover the handle ranges
on the fly, though we can also do that if it becomes necessary.