Increase the I2S player buffer size
Sometimes the sonar ping has artifacts in it when it plays at the same
time the launcher icons are redrawing. This is due to trying to read the
icons and music from the SD card at the same time. We can compensate for
this by doing a little more buffering in the I2S player module.