Fix a few nits
LEDs were being initialized twice in main.c.

Using a #define for time_t in datetime.h was causing a syntax error
with GCC 8.2.

With LTO turned off, the code would not build due to an undefined linker
reference to breakTime(). This function isn't implemented yet. With LTO,
the compiler/linker didn't actually emit a call to this function since
it could not be called unless BLACK_BADGE is #defined, and it knew that.
But without LTO, the compiler can't optimize out the reference. This
needs to be cleaned up.