Add NULL pointer protection
Since we have the MPU and we're already using to provide stack protection,
we may as well also use to to provide NULL pointer protection. The first
1024 bytes of the address space (0x00000000 to 0x000003FF are now marked
as read/write inhibited. Any code which tries to do a load or store to this
region will trigger an immediate memory manager fault. As it is now,
reading or writing to this region always succeeds (though writing has no
effect since flash is read only), and this can mask bugs where code tries
to read or write through a NULL pointer. I have a sneaking suspicion there
may be such a bug somewhere in the uGFX code and I'm hoping this will
help track it down.