Avoid UI setup on broken preview silicon
If we haven't initialized uGFX, we'll trigger a NULL pointer dereference
in all of the orchard code when we try to access the uninitialized GDISP
structure, so avoid launching the app thread.

Also, don't try to dispatch any orchard events from the joypad thread
either if the joypad event structure is uninitialized.