Fix some race conditions.
The newlib version of malloc requires locking, however by default the
__malloc_lock() and __malloc_unlock() functions are empty. It's up to
us to provide them, and previously I wasn't. This is bad, because since
ChibiOS is multithreaded, the possibility exists that we could have
malloc()s and/or free()s occuring in multiple threads simultaneously.
These functions are now implemented in newlib_syscalls.c using a ChibiOS
mutex to guard the heap.

Also, I discovered what I'm pretty sure is a bug in uGFX. There is
a Trigger_Redraw() macro which causes the gtimer thread to iterate over
all of the Gwin windows and perform any pending updates for them. The
The Gwin windows are held in a linked list which is protected by a mutex.
However the timer callback (RedrawTimerFn()) always calls
_gwinFlushRedraws() with REDRAW_NOWAIT. I think the logic is that it's
supposed to check if someone is holding the mutex and abort if it is
(because somebody else is already redrawing), but I'm not sure that
works correctly. I changed it to REDRAW_WAIT so that it always acquires
the mutex.

There's a slim chance that the uGFX crash I was seeing was a side effect
of the missing malloc()/free() locking, but I to be safe I'd rather make
both changes.