Add first attempt at notify app
Currently this app responds only when someone writes to the chat request
handle in the GATTS service. So far I've only tested this using my phone.
The chat app itself still needs to be updated to send the request.

I also added the alert1 effect from the DC25 badge design. We can replace
this with something else later. (Maybe a submarine alert kalxon?)