Try to add LED support to the firmware updater
The production board design has no status LEDs for the SD cars,
so we can't really see any progress indication when the firmware
updater is running. We can't really write to the screen since
we're too small to include graphics support. So to try and give
the user some feedback, we now try to use the LED chain. We can
light up all the LEDs red as an error indication, green for success,
and use a single marching blue LED as a progress indicator. For
the Nordic DK board, this code is basically a hardmless no-op
since there I2C bus pins are not connected to anything.

I'm crossing my fingers and hoping this works because John has
the only board with LEDs. I suppose I'll find out later if it
actually works.