Improve LED support in the updater
It seems that if you touch the exposed shitty add-on connector pads,
you can cause the I2C controller to lose communication with the LED
control chip. I'm still not sure how to recover correctly when this
happens. But I don't want the firmware updater to stall if someone
touches the pads while we're in the middle of flashing a new image
(since that will brick it, unless they have a JTAG probe). To mitigate
this, we now implement a dummy chThdSuspendTimeoutS() function that
actually times out after a while instead of always blocking forever,
and we now call the I2C TX function with a timeout instead of always
using TIME_INFINITE. This way if the LEDs get stuck, the firmware
update will still continue and complete successfully.