Recover the LEDs when the I2C bus glitches
If you touch the expose shitty add-on pads, you can glitch the I2C bus,
which causes the LEDs to freeze. We can detect when this happens if
the I2C transfer returns anything other than MSG_OK. (It will return
MSG_TIMEOUT when a transfer fails.) If this happens, we need to reset
the I2C controller and unjam the bus, then send some data through to
re-sync comms with the LED controller chip.

If someone keeps their fingers on the pads, it will keep glitch the
bus indefinitely and the LED updates will never go through, however
this won't block the flashing process entirely, so while the LEDs
won't flash, the update will still complete and the badge won't be
bricked.