Fix LED stalls
It turns out that to properly reset the I2C bus, you have to power-cycle
it using a magic, undocumented register at offset 0xFFC. To handle
LED stalls correctly every time, we have to disable the I2C controller,
power-cycle it, and then turn it back on again. We also have to make
sure not to use TIME_INFINITE when doing an I2C transaction (if there's
an error, we stall forever).