Implement airplane mode and sound on/off features in setup
Getting airplane more to work is tricky. To do it right, we have to
disable the radio. But when we disable the radio, we lose access to
the flash and the random number generator which are accessed via the
SoftDevice.

I finally bit the bullet and made the flash driver and random support
use either direct hardware access or the SoftDevice on the fly. You can
use sd_softdevice_is_enabled() to check if it's turned on, and if so
we use the SD APIs, otherwise we talk to the hardware directly.