Unbreak factory reset
I inadvertently broke this when I rearrange the start-up path when
I added airplane mode support.

Also, the path for the audio playback when a factory reset occurs
was wrong. There is no game/ping.snd file. There is a game/map_ping.snd
though. I'm assuming this is what was intended.

Also we need to pause for the reset sound effect to finish, otherwise it
gets garbled during radio startup. We may want to use a shorted tone
for this later.