Fix some sound paths
The unlocks app expects to use the Mortal Kombat "Pathetic" sound clip
to indicate failure, but we don't have that this time around. Replace it
with the Wilhelm scream until John decides on something better.

Also, for the factory recent, use the sonar ping instead of the game
map ping.