Add first attempt at a programming tool for the CP2102N
I finally figured out how to change the internal configuration of the
CP2102N. For right now, this program only resets the vendor and
product string to "Team Ides" and "DC27 Ides of DEF CON" respectively.
I still need to figure out how to make it change the GPIO pin setup.
I also need to make it more flexible so that it can do things like
dump the internal memory.

Unfortunately, while there are several chips in the CP210x product
line, they don't follow a consistent strategy for reading and writing
the internal configuration. The CP2102 and CP2102N use completely
different methods. The good news is that it's not actually possible
to lock the CP2102N. Care needs to be taken not to clobber the vendor
and device ID fields though.