Fix the GIF reader in uGFX
The code that parses extension blocks was written incorrectly. It
always read a few extra bytes so that it could parse Netscape looping
blocks (which is the only application-specific extension it knows how
to handle) which would leave the file read pointer out of position to
find the block length so that the code to skip to the next block correctly.