Fix a small I2C stability issue
When compiled without LTO, sometimes reads will return success even
if there's no device present on the I2C bus (such as with a Nordic
DK board with no LEDs). It looks like we need to force a few write
drains to the I2C registers in the interrupt handler to make sure
the driver and hardware stay in sync.