Update the OTA utility a little
Checksum the data as it's being sent so that the other end doesn't
try flashing a bogus firmware file.

Save the received file to OTA.BIN on the SD card instead of overwriting
the BADGE.BIN file directly. If the OTA.BIN file checksums correctly,
then we can replace BADGE.BIN with OTA.BIN.

Display a background image when sending or receiving an update.