Change the location of the remote authentication password
Instead of the temporary hardcoded string, we now use a string
stored in the user info area, which can be set using the provisioning
script when all the badges are prepared. This avoids having the
password stored in the source tree.

Of course anybody with access to the source could always just
build a custom firmware image without the password check...