Update the provisioning scripts
Use the flash write_bank command to write the firmware image. This is
faster: since we just did a mass erase command, which finishes right
away, we don't need to do another erase operation (which is what the
program command does).

Also, write the password in reverse-byte order. OpenOCD assumes we want
to write word values in little-endian mode, which swaps the bytes. This
turns "narf" info "fran".