Change the way we make icon backgrounds transparent
ImageMagick 7 seems to have a bug in it that breaks the conversion
command that I tried to use to make icon backgrounds transparent.
A couple of scripts that use an alternate, simpler approach have been
added which should work with both ImageMagick 6.9 and 7.

A script has also been added to just make backgrounds black. This
should allow us to still use the .rgb images for icons while still
having them appear non-square against the black background in the
launcher. The advantage to this is that while using transparent
.gifs works, the native .rgb images can be loaded faster.