Switch to using advanced kernel mode
I noticed that the music player stutters when it has the LED bargraph
and spectrograph going at the same time and the radio is on. (This is
only noticeable when playing fixed tones, but it's there.) Using
advanced kernel mode scheduling seems to dramatically improve this
(probably because in some cases it prevents interrupts from being
completely masked off in some places, in favor of adjusting the
CPU priority instead).

This should not have any effect on stability. I've tested this before
and it worked fine, but I decided to stick with compact kernel mode
just for simplicity's sake.

Also, I forgot to set the priority for the random number generator
block in mcuconf.h. For some reason this didn't show up in compact
mode, but it did now. Oops.